backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
